+++ packages/agent/src/certificate.test.ts
@@ -10,6 +10,7 @@ import { Principal } from '@dfinity/principal';
 import { decodeTime } from './utils/leb';
 import { readFileSync } from 'fs';
 import path from 'path';
+import { IC_ROOT_KEY } from './agent';
 
 function label(str: string): ArrayBuffer {
   return new TextEncoder().encode(str);
@@ -20,13 +21,6 @@ function pruned(str: string): ArrayBuffer {
 }
 
 (expect as any).addEqualityTesters([bufEquals]);
-
-// Root public key for the IC main net, encoded as hex
-const IC_ROOT_KEY =
-  '308182301d060d2b0601040182dc7c0503010201060c2b0601040182dc7c05030201036100814' +
-  'c0e6ec71fab583b08bd81373c255c3c371b2e84863c98a4f1e08b74235d14fb5d9c0cd546d968' +
-  '5f913a0c0b2cc5341583bf4b4392e467db96d65b9bb4cb717112f8472e0d5a4d14505ffd7484' +
-  'b01291091c5f87b98883463f98091a0baaae';
 
 test('hash tree', async () => {
   const cborEncode = fromHex(