+++ docs/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## [Unreleased]
 
+## [2.3.0] - 2025-02-07
+
+### Added
+
+- shouldFetchRootKey option added to `HttpAgent` constructor
+- ci: adds BOT_APPROVED_FILES config
+
+
+### Changed
+
+- feat: HttpAgent uses anonymous identity to make syncTime call, which can allow readState calls to work beyond 5 minutes
+- chore: bumps .nvmrc and nodejs version in CI to 22
+- HttpAgent now awaits fetching rootkey before making network calls if `shouldFetchRootKey` is set
+- chore: npm audit fixes
+- feat: enhanced details in agent call, query, and read_state errors
+  - error now includes hex encoded response, requestId, sender_pubkey, and sender_sig in addition to message for improved debugging process
+
 ## [2.2.1] - 2025-02-07
 
 ### Changed